split out files
