+++ src/editing/PageView.tsx
@@ -1,13 +1,5 @@
-import React, {
-    CSSProperties,
-    MouseEvent,
-    ReactNode,
-    useContext,
-    useEffect,
-    useRef,
-    useState
-} from "react";
-import {Bounds, Point, Size} from "josh_js_util";
+import React, {MouseEvent, useContext, useEffect, useRef, useState} from "react";
+import {Point, Size} from "josh_js_util";
 import {HBox, PopupContext} from "josh_react_util";
 import {GlobalState} from "../models/state";
 import {
@@ -15,10 +7,8 @@ import {
     DrawableClass,
     DrawableShape,
     FamilyPropChanged,
-    Handle,
     PageClass,
-    PropChanged,
-    ScaledSurface
+    PropChanged
 } from "../models/om";
 import {MenuActionButton, MenuBox, useObjectProxyChange, useObservableChange} from "../common";
 import {
@@ -41,114 +31,10 @@ import {EditState, PathShapeEditHandler} from "./PathShapeEditHandler"
 import {DragHandler} from "./DragHandler"
 import "./PageView.css"
 import {NGonClass} from "../models/ngon";
-import {lookup_dpi, point_to_pixels, Unit} from "../models/unit";
+import {lookup_dpi, Unit} from "../models/unit";
+import {ScaledDrawingSurface} from "./scaled_drawing";
+import {FloatingPalette} from "./FloatingPalette";
 
-
-export class ScaledDrawingSurface implements ScaledSurface {
-    private ctx: CanvasRenderingContext2D;
-    private scale: number;
-    private unit: Unit;
-
-    constructor(ctx: CanvasRenderingContext2D, scale: number, unit: Unit) {
-        this.ctx = ctx
-        this.scale = scale
-        this.unit = unit
-    }
-    fillRect(bounds: Bounds, fill: "string") {
-        this.ctx.fillStyle = fill
-        this.ctx.fillRect(bounds.x*this.scale,bounds.y*this.scale,bounds.w*this.scale,bounds.h*this.scale)
-    }
-    outlineRect(bounds: Bounds) {
-        this.ctx.strokeRect(bounds.x*this.scale, bounds.y*this.scale, bounds.w*this.scale, bounds.h*this.scale)
-    }
-    strokeRect(bounds: Bounds, strokeFill: string, strokeWidth: number) {
-        this.ctx.strokeStyle = strokeFill
-        this.ctx.lineWidth = strokeWidth
-        this.ctx.strokeRect(bounds.x*this.scale, bounds.y*this.scale, bounds.w*this.scale, bounds.h*this.scale)
-    }
-
-    fillRoundRect(bounds: Bounds, radius: number, fill: any) {
-        this.ctx.fillStyle = fill
-        this.ctx.beginPath()
-        this.ctx.roundRect(bounds.left()*this.scale, bounds.top()*this.scale, bounds.w*this.scale, bounds.h*this.scale, radius)
-        this.ctx.closePath()
-        this.ctx.fill()
-    }
-    strokeRoundRect(bounds: Bounds, radius: number, strokeFill: string, strokeWidth: number) {
-        this.ctx.strokeStyle = strokeFill
-        this.ctx.lineWidth = strokeWidth
-        this.ctx.beginPath()
-        this.ctx.roundRect(bounds.left()*this.scale, bounds.top()*this.scale, bounds.w*this.scale, bounds.h*this.scale, radius)
-        this.ctx.closePath()
-        this.ctx.stroke()
-    }
-
-    fillArc(center: Point, radius: number, startAngle: number, endAngle: number, fill: string) {
-        this.ctx.fillStyle = fill
-        this.ctx.beginPath()
-        this.ctx.arc(center.x*this.scale, center.y*this.scale, radius*this.scale, startAngle, endAngle)
-        this.ctx.fill()
-    }
-    outlineArc(center: Point, radius: number, startAngle: number, endAngle: number, fill: string) {
-        this.ctx.beginPath()
-        this.ctx.arc(center.x*this.scale, center.y*this.scale, radius*this.scale, startAngle, endAngle)
-        this.ctx.stroke()
-    }
-
-    private calcFont(fontSize:number) {
-        return `${fontSize}pt sans-serif`
-    }
-    fillText(text: string, center: Point, fill: string, fontSize: number) {
-        this.ctx.fillStyle = fill
-        this.ctx.font = this.calcFont(fontSize)
-        this.ctx.fillText(text, center.x*this.scale, center.y*this.scale)
-    }
-    fillLinePath(position: Point, points: Point[], closed: boolean, filled: boolean, fill: string) {
-        if(points.length < 3) return
-        this.ctx.save()
-        this.ctx.fillStyle = fill
-        this.ctx.translate(position.x*this.scale, position.y*this.scale)
-        this.ctx.beginPath()
-        this.ctx.moveTo(points[0].x*this.scale,points[0].y*this.scale)
-        for (let pt of points) this.ctx.lineTo(pt.x*this.scale, pt.y*this.scale)
-        if(closed) this.ctx.closePath()
-        this.ctx.fill()
-        this.ctx.restore()
-    }
-    outlineLinePath(position: Point, points: Point[]) {
-        if(points.length < 3) return
-        this.ctx.save()
-        this.ctx.translate(position.x*this.scale, position.y*this.scale)
-        this.ctx.beginPath()
-        this.ctx.moveTo(points[0].x*this.scale,points[0].y*this.scale)
-        for (let pt of points) this.ctx.lineTo(pt.x*this.scale, pt.y*this.scale)
-        this.ctx.closePath()
-        this.ctx.stroke()
-        this.ctx.restore()
-    }
-
-    overlayHandle(h: Handle) {
-        this.ctx.fillStyle = 'red'
-        let p = point_to_pixels(h.getPosition(),this.unit)
-        this.ctx.fillRect(p.x-10,p.y-10,20,20)
-    }
-    dragRect(dragRect: Bounds) {
-        this.ctx.strokeStyle = 'cyan'
-        this.ctx.lineWidth = 1
-        this.ctx.strokeRect(dragRect.x*this.scale, dragRect.y*this.scale, dragRect.w*this.scale, dragRect.h*this.scale)
-    }
-
-    overlayFillText(s: string, point: Point) {
-        this.ctx.fillStyle = 'black'
-        this.ctx.font = '20pt sans-serif'
-        this.ctx.fillText(s, point.x, point.y)
-    }
-    overlayPoint(point: Point, green: string) {
-        this.ctx.fillStyle = green
-        let r = 5
-        this.ctx.fillRect(point.x*this.scale-r, point.y*this.scale-r, r*2, r*2)
-    }
-}
 
 function drawCanvas(canvas: HTMLCanvasElement, page: PageClass, doc: DocClass, state: GlobalState, handler:MouseHandlerProtocol, scale:number) {
     let ctx = canvas.getContext('2d') as CanvasRenderingContext2D
@@ -166,7 +52,6 @@ function drawCanvas(canvas: HTMLCanvasElement, page: PageClass, doc: DocClass, s
         ctx.lineWidth = 10;
         if (sel instanceof DrawableClass) sel.drawSelected(surf);
     }
-    handler.drawOverlay(surf,state)
     //draw the handles
     for(let sel of selected) {
         if (sel instanceof DrawableClass) {
@@ -174,35 +59,8 @@ function drawCanvas(canvas: HTMLCanvasElement, page: PageClass, doc: DocClass, s
             if(h) surf.overlayHandle(h)
         }
     }
+    handler.drawOverlay(surf,state)
     ctx.restore()
-}
-
-function FloatingPalette(props: { children: ReactNode, visible:boolean }) {
-    const [position, setPosition] = useState(new Point(0,0))
-    const style:CSSProperties = {
-        top: `${position.y}px`,
-        left: `${position.x}px`,
-        visibility: props.visible?'visible':'hidden',
-    }
-    const mouseDown = (e:MouseEvent<HTMLDivElement>) => {
-        let pos = position
-        const dragger = (e:any) => {
-            let trans = new Point(e.movementX,e.movementY)
-            pos = pos.add(trans)
-            setPosition(pos)
-        }
-        const upper = (e:any) => {
-            window.removeEventListener('mousemove',dragger)
-            window.removeEventListener('mouseup',upper)
-        }
-        window.addEventListener('mousemove', dragger)
-        window.addEventListener('mouseup',upper)
-    }
-    return <div className={'floating-palette'} style={style}
-                onMouseDown={mouseDown}>
-        <header>tools</header>
-        {props.children}
-    </div>
 }
 
 export function PageView(props:{doc:DocClass, page:PageClass, state:GlobalState}) {
